requirements.txt
discord.py>=2.5.0
httpx>=0.28.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
